requirements.txt
python-telegram-bot==13.15
gspread==5.11.3
oauth2client==4.1.3
six>=1.15.0
urllib3>=1.26.0
